requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.3
python-docx==1.1.2
python-dotenv==1.0.1
sse-starlette==2.1.3
